Version 1.2.12 Clean up on definitive uninstall only (or else some stuff from the new package would also be erased).

+++ emu/drivers/emu-e1000e_emu.spec
@@ -4,7 +4,7 @@
 
 Summary: CMS Emu local DAQ Gbit and peripheral crate VME drivers for kernel %{kernel_version} based on the e1000e module for the Intel dual port NIC model PRO/1000 PF
 Name: emu-e1000e_emu
-Version: 1.2.11
+Version: 1.2.12
 Release: 1.slc6
 License: none
 Group: none
@@ -47,20 +47,7 @@ touch %{_topdir}/BUILD/MAINTAINER
 
 %files
 %defattr(744,root,root,-)
-/usr/local/bin/e1000e_emu/eth_hook_2_ddu.ko
-/usr/local/bin/e1000e_emu/eth_hook_3_ddu.ko
-/usr/local/bin/e1000e_emu/eth_hook_4_ddu.ko
-/usr/local/bin/e1000e_emu/eth_hook_5_ddu.ko
-/usr/local/bin/e1000e_emu/eth_hook_2_dmb.ko
-/usr/local/bin/e1000e_emu/eth_hook_3_dmb.ko
-/usr/local/bin/e1000e_emu/eth_hook_4_dmb.ko
-/usr/local/bin/e1000e_emu/eth_hook_5_dmb.ko
-/usr/local/bin/e1000e_emu/eth_hook_2_vme.ko
-/usr/local/bin/e1000e_emu/eth_hook_3_vme.ko
-/usr/local/bin/e1000e_emu/eth_hook_4_vme.ko
-/usr/local/bin/e1000e_emu/eth_hook_5_vme.ko
-/usr/local/bin/e1000e_emu/e1000e_emu.ko
-/usr/local/bin/e1000e_emu/load_e1000e_emu.sh
+/usr/local/bin/e1000e_emu
 # Files required by Quattor
 %defattr(644,root,root,755)
 %doc MAINTAINER ChangeLog README
@@ -74,16 +61,20 @@ echo "[[ -x /usr/local/bin/e1000e_emu/load_e1000e_emu.sh ]] && /usr/local/bin/e1
 /usr/local/bin/e1000e_emu/load_e1000e_emu.sh || true
 
 %preun
-# Unload modules
-[[ $(/sbin/lsmod | grep -c e1000e_emu) -eq 0 ]] || /sbin/modprobe -r e1000e_emu || true
+if [[ $1 -eq 0 ]]; then
+   echo "Definitive uninstall of this package. Cleaning up."
 
-# Stop loading daq drivers at boot time.
-[[ -f /etc/rc.d/rc.local ]] && sed -i -e "/\/usr\/local\/bin\/e1000e_emu\/load_e1000e_emu.sh/d" /etc/rc.d/rc.local || true
+   # Unload modules
+   [[ $(/sbin/lsmod | grep -c e1000e_emu) -eq 0 ]] || /sbin/modprobe -r e1000e_emu || true
 
+   # Remove modules from /lib/modules
+   rm -f /lib/modules/%{kernel_version}/kernel/drivers/net/e1000e/e1000e_emu.ko || true
+   rm -f /lib/modules/%{kernel_version}/kernel/drivers/net/e1000e/eth_hook_*.ko || true
+
+   # Update module dependencies
+   /sbin/depmod -a || true
+
+   # Stop loading daq drivers at boot time.
+   [[ -f /etc/rc.d/rc.local ]] && sed -i -e "/\/usr\/local\/bin\/e1000e_emu\/load_e1000e_emu.sh/d" /etc/rc.d/rc.local || true
+fi
 %postun
-# Remove modules from /lib/modules
-rm -f /lib/modules/%{kernel_version}/kernel/drivers/net/e1000e/e1000e_emu.ko || true
-rm -f /lib/modules/%{kernel_version}/kernel/drivers/net/e1000e/eth_hook_*.ko || true
-rm -rf /usr/local/bin/e1000e_emu || true
-# Update module dependencies
-/sbin/depmod -a || true
